connect app to local mongodb
