+++ server.js
@@ -1,22 +1,11 @@
-var express = require('express');
-var bodyParser = require('body-parser');
-var cors = require('cors');
-var path = require('path');
-
-var app = express();
-
-// add middleware
-app.use(cors());
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({
-    extended:false 
-}));
+// load env variables
+if (process.env.NODE_ENV !== 'production') {
+	require('dotenv').config();
+}
 
 const port = 3000;
 
-app.use('/', (req, res) => {
-  res.send('SMS Management API');
-});
+const app = require('./src/app');
 
 app.listen(port, function(req, res) {
   console.log('Server started at port: ', port);
